Make extinguishers work on cigarettes too
A bunch of renaming to make the rest of my code work with SmokableComponent
