+++ Content.Server/Atmos/EntitySystems/FlammableSystem.cs
@@ -325,7 +325,7 @@ namespace Content.Server.Atmos.EntitySystems
 
             _ignitionSourceSystem.SetIgnited(uid, false);
 
-            var extinguished = new FlammableExtinguished();
+            var extinguished = new ExtinguishedEvent();
             RaiseLocalEvent(uid, ref extinguished);
 
             UpdateAppearance(uid, flammable);
@@ -350,7 +350,7 @@ namespace Content.Server.Atmos.EntitySystems
                     _adminLogger.Add(LogType.Flammable, $"{ToPrettyString(uid):target} set on fire by {ToPrettyString(ignitionSource):actor}");
                 flammable.OnFire = true;
 
-                var extinguished = new FlammableIgnited();
+                var extinguished = new IgnitedEvent();
                 RaiseLocalEvent(uid, ref extinguished);
             }
