+++ Content.Shared/Atmos/Components/ExtinguishableSetCollisionWakeComponent.cs
@@ -3,9 +3,9 @@
 namespace Content.Shared.Atmos.Components;
 
 /// <summary>
-/// Makes entities with <c>flammable</c>component automatically enable/disable <see cref="CollisionWakeComponent"/>,
+/// Makes entities with extinguishing behavior automatically enable/disable <see cref="CollisionWakeComponent"/>,
 /// so they can be extinguished with fire extinguishers.
 /// </summary>
 [RegisterComponent]
 [NetworkedComponent]
-public sealed partial class FlammableSetCollisionWakeComponent : Component;
+public sealed partial class ExtinguishableSetCollisionWakeComponent : Component;
