+++ Resources/Prototypes/Entities/Objects/Misc/candles.yml
@@ -27,7 +27,7 @@
           variation: 0.05
           volume: 10
     - type: UseDelay
-    - type: FlammableSetCollisionWake
+    - type: ExtinguishableSetCollisionWake
     - type: Flammable
       fireSpread: false
       canResistFire: false
